Moved QtTimer implementation to glue.backends, and don’t automatically run get_qapp anymore (just make sure it’s always run when needed, for example when running tests or initializing the GlueApplication)
