+++ glue/conftest.py
@@ -25,6 +25,10 @@ def pytest_configure(config):
     from glue import config
     config.CFG_DIR = tempfile.mkdtemp()
 
+    # Start up QApplication
+    from glue.external.qt import get_qapp
+    app = get_qapp()
+
     # Force loading of plugins
     from glue.main import load_plugins
     load_plugins()
