+++ glue/utils/qt/tests/test_qmessagebox_widget.py
@@ -1,6 +1,6 @@
 from __future__ import absolute_import, division, print_function
 
-from glue.qt import get_qapp
+from glue.external.qt import get_qapp
 
 from .. import QMessageBoxPatched as QMessageBox
